add Table of requests

+++ kickstart/ethereum/contracts/Campaign.sol
@@ -91,7 +91,7 @@ contract Campaign {
         );
     }
 
-    function getRequestsCount() public view returns (uint) {
+    function getRequestsCount() public view returns (uint256) {
         return requests.length;
     }
